Move endpoints into separate base class. Sort them alphabetically.
Next major step into breaking down the monolith, and clearing the
__init__.py pollution.
